backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.3
aiohttp==3.9.3
python-dotenv==1.0.1 